Add Action/Cancel NavRequest support
This requires introducing the concept of `focus_parent` in `NavFence`.
The focus_parent is the `Focusable` that "leads" to the `NavFence`.

Also introduces the concept of `Dormant` focusable to support "focus
memory".

I also added a lot of documentation and the `nested_menu_1d_nav` example
to show off how nested navigation works.