Add tabbed navigation with example
* Replace `NavRequest::{Next, Previous}` by `NavRequest::MenuMove`
* Implement logic for tabbed navigation

Note: the "ultimate navigation example" uses my macros to make the UI.
I use it because I find it prettier than the non-macro version. I think
it is also easier to follow the child/parent relationships of entities
with this macro. I think the noise of the "normal" way of declaring UI
makes it difficult to read the declarition and understand it.