Add loops/wrapping navigation
Add ability to set NavFences as looping. Within a looping NavFence, when
going leftward when the leftmost element is focused, go to the rightmost
element (while keeping the closest possible to the X/Y axis of the
focused element, when relevant)

This also applies to Sequence menus, although the ordering is based on
index in the element list rather than the geometric positioning.

Add some loops (brother) to the examples to give a feel.