Fix cleanup of dormant nodes with mouse navigation
By adding the `non_inert_child` field to NavMenu, it is now possibly to
directly disactivate previously non-inert entries in a menu whenever we
update which one is to be the non-inert one.

We also probably make the `Action` request a bit more performant, by
directly accessing the non-inert child rather than collecting all the
focusable children of a menu and then querying which one is non-inert in
that collection.