Fix 2d wrapping behavior
The previous way of handling wrapping was very weird. Now we do it in
a more intuitive manner. ie: when we reach the border of the screen in
one direction and want to go in that direction, we resolve the move
request as if we were focused on an element off-screen in the other
direction.