Add marking menu bundle
This resolves #4, see rationale[1]. The idea is to simplify identifying
Focusable entities as being part of a specific NavMenu. By default it's
not possible. But using the code introduced by this commit, the user
will have the possibility to hook into the navigation system to add
markers that propagate to Focusables within menus.

[1] https://github.com/nicopap/ui-navigation/issues/4#issuecomment-1019140598