Fix marking systems panicking at startup
A change in bevy made `Added<Foo>` in filter position conflict with
`&mut Foo` in query position, for soundness. See [bug-report].

This is a sound change, however, it made the marking system `Added`
parameters incompatible with the `NavQueries` parameter. I fixed it by
removing the `&mut` queries from the `NavQueries` struct, and using a
`ParamSet` in the `listen_nav_request` system.

This fixes #21.

[bug-report]: https://github.com/bevyengine/bevy/issues/4624