Fix Cycling behavior
Previously, the cycling method would "wrap" the cursor by setting the
relevant axis value to a sufficiently high value. This was brittle, as
arbitrary UI camera perspective and positioning would break that
heuristic.

Now, it accounts for those variables, and works in all cases where the
entirety of the UI is visible on screen.

This fixes #17.