Fix default_mouse_input not accounting for scale
This fixes an issue where the UI camera scale would break the mouse
input system.

This is a breaking change, since we also remove the undocumented
`UiCamera` marker struct. Since now bevy has the `CameraUi` component,
that marker struct is not needed anymore.

We also add a cool example to demonstrate handling of UI camera scale
and offset, and how cycling works.