Add a stress test example
In the optics of using an acceleration data structure for mouse picking
and navigation, I added a test example with a very large amount of
focusables.

It turns out the navigation systems do not eat up CPU, at least, so
little compared to the bevy UI stuff that it's not worth optimizing.
Since it ends up being such a tinny fraction of the overall performance
bill, optimizing them would be worthless.