Fix README and crate-level documentation links
We use markdown link references to have different links in the rustdoc
documentation and the github README.md. This way, when reading locally
the documentation, we get the links to the internal docs rather than
links to docs.rs. This also will make it easier to update links in the
future.