Add ScreenBoundaries resource for wrapping
This makes bevy_ui_navigation yet again independent from bevy_ui, which
is required for the `bevy-ui` feature flag to work.

This fixes #23.

We also add a `DefaultNavigationPlugins` plugin group to make it even
easier to get started using the crate. This is because the introduction
of `ScreenBoundaries` required adding a new "user" system:
`systems::update_boundaries`. A side effect is that it's much cleaner to
use this crate, while keeping the flexibility it started with.