Remove focus follow mouse in default mouse input
Focus following mouse cursor is contrary to the way most UI frameworks
handle focus. This as the default behavior is surprising. Furthermore,
the default systems do not provide a more traditional mouse focus
handler.

This commit adds a way to have a more traditional mouse focus system and
makes it the default.

Now, focus changes to the hovered element only when the mouse button is
pressed, or when it is released if the currently hovered element wasn't
focused yet.

We also add a `focus_follows_mouse` field to the `InputMapping` to allow
users to switch to the older behavior.

This also makes it much easier to handle situations when we click on a
button that changes focus.

A major downside is that the user must now implement themselves hover
highlight.

Fixes #33

- https://github.com/nicopap/ui-navigation/issues/33