Replace mouse input by bevy_mod_picking
We now rely entirely on the accuracy of the bevy_mod_picking
implementation for the pointer-based focus system.

Instead of having our extremely hacky impl, we rely on bevy_mod_picking,
which fixes a number of issues, notably it fixes #3.

This also allows users to more easily integrate with other kind of
focusables.