Add CI and pre-hook script
This makes PRs easier to manage, as the tests are automatically ran.

It also makes it safer to commit, as the `pre-hook` rule can be used as
a pre-commit hook to prevent commiting code that doesn't work.

I also marked the Menu action example as no_run, since it doesn't really
help to run it, and it makes tests a lot slower.